successfully pulled the tox_specific items into 	ox_harness.py. time to start getting the pooling working
